refactor: move BigQuery parser into format dispatch system
Moves BigQuery CTE parsing from a dialect-branch hack into a proper format
module under src/formats/, registered in parse.js alongside Prisma, Mermaid,
etc. Fixes a bug where selecting BigQuery dialect silently overrode the
Format dropdown. Also makes comment-stripping quote-aware to match parser.js's
existing behaviour.
