+++ src/parse.js
@@ -1,6 +1,7 @@
 // Input-format dispatcher. SQL is the editable core; Prisma / SQLAlchemy /
 // Sequelize are parse-only. Auto-detects from the text unless a format is forced.
 import { parseSchema as parseSQL } from './parser.js';
+import { parseBigQuery } from './formats/bigquery.js';
 import { parsePrisma } from './formats/prisma.js';
 import { parseSQLAlchemy } from './formats/sqlalchemy.js';
 import { parseSequelize } from './formats/sequelize.js';
@@ -11,6 +12,7 @@ import { parsePlantUML } from './formats/plantuml.js';
 export const FORMATS = {
   auto: 'Auto-detect',
   sql: 'SQL',
+  bigquery: 'BigQuery',
   prisma: 'Prisma',
   dbml: 'DBML',
   mermaid: 'Mermaid',
@@ -27,6 +29,7 @@ export function detectFormat(text) {
   if (/^\s*model\s+\w+\s*\{/m.test(t) || /\b(datasource|generator)\s+\w+\s*\{/.test(t)) return 'prisma';
   if (/\.define\s*\(\s*['"]/.test(t) || (/DataTypes\./.test(t) && /\.init\s*\(/.test(t))) return 'sequelize';
   if (/\bColumn\s*\(|mapped_column\s*\(/.test(t) && /\b(Base|db\.Model|DeclarativeBase|declarative_base|__tablename__)\b/.test(t)) return 'sqlalchemy';
+  if (/^\s*with\b/i.test(t) && /`[^`]+\.[^`]+`/.test(t)) return 'bigquery';
   return 'sql';
 }
 
@@ -35,6 +38,7 @@ export function parseSchema(text, format = 'auto') {
   let res;
   try {
     switch (fmt) {
+      case 'bigquery': res = parseBigQuery(text); break;
       case 'prisma': res = parsePrisma(text); break;
       case 'dbml': res = parseDBML(text); break;
       case 'mermaid': res = parseMermaid(text); break;
